pyproject.toml
[tool.poetry]
name = "easycrm"
version = "0.1.0"
description = ""
authors = ["Markus Wahlman <markus.wahlman@helsinki.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.1"
autopep8 = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
